+++ lilya/middleware/clickjacking.py
@@ -45,11 +45,10 @@ class XFrameOptionsMiddleware(MiddlewareProtocol):
         """
         if message["type"] == "http.response.start":
             # we need to update the message
-            headers: Header = Header.from_scope(message)
+            headers: Header = Header.ensure_header_instance(message)
 
             if headers.get("X-Frame-Options") is None:
                 headers.add("X-Frame-Options", self.get_xframe_options())
-                message["headers"] = headers.encoded_multi_items()
         await send(message)
 
     def get_xframe_options(self) -> str: