+++ lilya/middleware/cors.py
@@ -95,7 +95,7 @@ class CORSMiddleware(MiddlewareProtocol):
             return
 
         method = scope["method"]
-        headers = Header.from_scope(scope=scope)
+        headers = Header.ensure_header_instance(scope=scope)
         origin = headers.get("origin")
 
         if origin is None:
@@ -241,9 +241,8 @@ class CORSMiddleware(MiddlewareProtocol):
             await send(message)
             return
 
-        message.setdefault("headers", [])
         # we need to update the message
-        headers = Header.from_scope(scope=message)
+        headers = Header.ensure_header_instance(scope=message)
         headers.update(self.simple_headers)
         origin = request_headers["Origin"]
         has_cookie = "cookie" in request_headers
@@ -253,7 +252,6 @@ class CORSMiddleware(MiddlewareProtocol):
         elif not self.allow_all_origins and self.validate_origin(origin=origin):
             self.set_explicit_origin(headers, origin)
 
-        message["headers"] = headers.encoded_multi_items()
         await send(message)
 
     @staticmethod