+++ lilya/middleware/security.py
@@ -124,7 +124,7 @@ class SecurityMiddleware(MiddlewareProtocol):
         """
         if message["type"] == "http.response.start":
             # we need to update the headers of message
-            headers: Header = Header.from_scope(message)
+            headers: Header = Header.ensure_header_instance(message)
             headers.add(
                 "Content-Security-Policy", "" if not self.content_policy else self.content_policy
             )
@@ -134,6 +134,5 @@ class SecurityMiddleware(MiddlewareProtocol):
             headers.add("X-Content-Type-Options", self.content_type_options)
             headers.add("X-Frame-Options", self.frame_options)
             headers.add("X-XSS-Protection", self.xss_protection)
-            message["headers"] = headers.encoded_multi_items()
 
         await send(message)