+++ lilya/routing.py
@@ -791,7 +791,7 @@ class Host(BasePath):
             Tuple[Match, Scope]: The match result and child scope.
         """
         if scope["type"] in {ScopeType.HTTP, ScopeType.WEBSOCKET}:
-            headers = Header.from_scope(scope=scope)
+            headers = Header.ensure_header_instance(scope=scope)
             host = headers.get("host", "").split(":")[0]
             match = self.host_regex.match(host)
             if match:
@@ -1951,7 +1951,7 @@ class Include(BasePath):
             key: self.param_convertors[key].transform(value)
             for key, value in match.groupdict().items()
         }
-        remaining_path = f'/{matched_params.pop("path", "")}'
+        remaining_path = f"/{matched_params.pop('path', '')}"
         matched_path = route_path[: -len(remaining_path)]
 
         path_params = {**scope.get("path_params", {}), **matched_params}